duplicates of models.py resolved. routes for autentification require rework.
