--- app/src/database/models.py
@@ -1,39 +1,0 @@
-from sqlalchemy.orm import relationship, Mapped, mapped_column
-from sqlalchemy.sql.schema import ForeignKey
-from sqlalchemy.sql.sqltypes import DateTime
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String, Date, Boolean, func
-
-
-Base = declarative_base()
-
-class User(Base):
-    __tablename__ = "users"
-    id = Column(Integer, primary_key=True)
-    username = Column(String(50))
-    email = Column(String(250), nullable=False, unique=True)
-    password = Column(String(255), nullable=False)
-    created_at = Column("created_at", DateTime, default=func.now())
-    avatar = Column(String(255), nullable=True)
-    refresh_token = Column(String(255), nullable=True)
-    confirmed = Column(Boolean, default=False)
-
-class BlacklistToken(Base):
-    __tablename__ = "blacklist_tokens"
-    id = Column(Integer, primary_key=True, index=True)
-    token = Column(String(255), nullable=False, unique=True)
-    
-class Comment(Base):
-    __tablename__ = "comments"
-    id: Mapped[int] = mapped_column(primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
-    photo_id: Mapped[int] = mapped_column(Integer, ForeignKey("photos.id"), nullable=False)
-    text: Mapped[str] = mapped_column(String(500), nullable=False)
-    created_at: Mapped[DateTime] = mapped_column(DateTime, default=func.now())
-    updated_at: Mapped[DateTime] = mapped_column(DateTime, default=func.now(), onupdate=func.now())
-    user: Mapped["User"] = relationship("User", backref="comments", lazy="joined")
-    photo: Mapped["Photo"] = relationship("Photo", backref="comments", lazy="joined")
-
-class Photo(Base):
-    __tablename__ = "photos"
-    id: Mapped[int] = mapped_column(primary_key=True)
